Changed the Threadly interfaces such that they will all return ListenableFuture's over the more basic Future.  Since ListernableFuture's extend Future, this is a backwards compatible change.
This allows for some simplification, as well as some code reduction.  It would only be a problem if someone wanted to implement one of the threadly interfaces, but did not want to use ListernableFuture's (of course I don't understand why that would be the case, so I am just going with the simpler and less code option).