Tried to acomplish several things.  Instead of setting everything up and destroying it on every TEST, we tried to do BeforeClass and AfterClass for some things which can be (safely) shared between tests.  These shared items are primarily thread pools, if they start to interact, we will revisit and see how we can fix that.
In addition I cleaned up how we were setting and handling the UncaughtExceptionHandler within tests.  I still don't want uncessary logs to be printed when they are expected, but we were setting it all over the place.  It is now only being set in test classes where I know there could be a false log message.