Reduces some of the complexity around the automation stop
This removes the extra thread started by the profiler, as well as the checking if the provided executor is a scheduler.
Instead we just assign variable which can be ran every loop of collection stats.  It checks the time and stops when needed.
One key aspect of this is that we have to remove the reference to the runnable after the stop is issued, otherwise the Profiler will never garbage collect.