Ability to cancel collection futures in FutureUtils
This is a new feature, previously it was documented that these futures could not be canceled.
That technically changed in 4.0.0 when SLF added the ability to cancel (so .cancel calls would return true).
Despite returning true it would not actually cancel anything.  This makes it so that cancel cascades to all futures which have not completed yet.