Fix for RateLimiterExecutor limiting under limit
This fixes a condition where the RateLimiterExecutor would limit a fraction of a second despite not reaching the limit yet.
The previous behavior would still enforce a limit, but too aggressively compared to what most would expect.
This change makes the behavior such that there will be no limiting, until that permits/per second is reached.  Once that is reached everything is scheduled out 1 second + amount of time for previous rate allocation to expire.
What that means is in the previous implementation rate limiting was smooth, in the new implementation if tasks finish faster than 1 second, they are going to be ran much faster until a limit is reached.

This change also resolves #153 by using a double internally to handle factions of a millisecond limitations.