Revert "Fix for RateLimiterExecutor limiting under limit"
This reverts commit 990e38926a38795c88a5d4a61692695998838467.

The more I thought about this, this new behavior does not make sense.
This makes the execution too bursty.  The previous behavior makes sense when you think in the sub-second.
This makes the behavior return to where you must pay the schedule penalty for the sub-max permit cost that was provided.