Interface rename: Dropped the "Interface" suffix
Originally I had put "Interface" at the end of all interface names so it was obvious what they were (with the exception of ListenableFuture which has no such convention).
The goal was to make it easy when looking through the javadocs to know what was an implementation, vs an interface.
I am now rethinking that, and wondering if it is just adding a bunch of extra characters/typing that don't really help in understanding.

This changes:
CallableContainerInterface -> CallableContainer
RunnableContainerInterface -> RunnableContainer
PrioritySchedulerInterface -> PrioritySchedulerService
SchedulerServiceInterface -> SchedulerService
SimpleSchedulerInterface -> SimpleScheduler
SubmitterExecutorInterface -> SubmitterExecutor
SubmitterSchedulerInterface -> SubmitterScheduler
ExceptionHandlerInterface -> ExceptionHandler

Right now this is fairly awkward to maintain backwards compatibility.  The old interfaces extend the new.  Where we accept implementations we accept the new interface (and since old extends new, this should be fine).  Where we implement an interface, or return an interface, we still return the old.  This is to ensure the old interface definitions would be assignable.