Reduced code duplication between NoThreadScheduler and PriorityScheduler
This is a start towards issue #159 to make NoThreadScheduler also be a PrioritySchedulerService.
While NoThreadScheduler will implement the priority interface with this change, priority is currently ignored.
I wanted to break this change up a bit so that it could be easier to digest.
With this change both NoThreadScheduler and PriorityScheduler extend "AbstractPriorityScheduler".  This class is mostly designed around PriorityScheduler, forcing NoThreadScheduler to extend and override implementation as necessary for it.

This does change the logic slightly inside NoThreadScheduler, most notably:
* There is no longer a concept of "isRunning" for a task.  This was used as a spot check for when clearing the queue, and for when checking if there are tasks ready to run.  This check may have _helped_ in avoiding returning tasks from shutdownNow() which actually executed.  But it did not garuntee it, and since PriorityScheduler did not need it, for now I would prefer to remove the concept.
* The task queue lifecycle for NoThreadScheduler changed...I wanted to match PriorityScheduler as close as possible since it has had the most benchmarking, but there may be room for improvement here.