Added "getScheduledTaskCount" to "SchedulerService" interface
This functionality should be easy for every reasonable scheduler to implement, so it was moved to the interface level.
The version which takes a priority is not in the PrioritySchedulerService interface because it is not easy to implement unless queues are maintained seperately (which is not the case for the limiter).  So for now that just moved into AbstractPriorityScheduler.
This combined with changes introduced in PR #165 resolves #157