Deprecated PrioritySchedulerLimiter
It does not make sense to limit mixed priorities how currently implemented.
The issue is that all priorities are queued equally.
Current assume that a PriorityScheduler is maxed out, and a PrioritySchedulerLimiter is actively limiting into it.  Say it submitted a Low priority task into the scheduler, and was delayed only because of it being low priority.  A high priority task may be waiting inside the limiter waiting for the low priority to finish.

A better alternative would be to wrap a PriorityScheduler in a PrioritySchedulerWrapper to set a default priority, then use a SchedulerServiceLimiter.