Run unit tests in parallel!
We have so many unit tests that it's time to speed this up by running them concurrently.
We can't run one fork per core because several unit tests end up using more than 1 core.  This causes issues on our slow jenkins build slaves in EC2 if we crank the concurrency up so high.
So this really only will be faster on systems with > 4 cores.  If anyone has a better idea of how to detect what parallism we can run, let me know!