Add "Promise" as a short hand to "ListenableFuture"
This comes from #180, with a goal to have less typing in this very frequently used structure.
I am a little nervous to completely remove ListenableFuture.  While I understand threadly has long names that is to make it easy to navigate the project through javadocs.
This is an attempt to provide both options, but I am unsure if it fits very cleanly.  Here we create a new interface "Promise" which just extends "ListenableFuture".  We now return "Promise" as an alternative, but if your setting it into a ListenableFuture it should work just as good.  We are not deprecating ListenableFuture, just giving a short hand version.

If we do like this, some outstanding questions are:
* What should we do about the interface "ListenableRunnableFuture"?  Should we have an extending "RunnablePromise" to keep things consistent?
* Should we rename "ListenableFutureTask" into "PromiseTask" to keep things shorter and consistent?