RateLimiterExecutor: Remove locking and do compare and swap instead
This removes the "rateLock" inside the "RateLimiterExecutor", and instead does optimistic CAS operations.
Advantages:
  * Obviously no locking
Disadvantages:
  * Requires going to native code each time we call "Double.doubleToLongBits" and "Double.longBitesToDouble"
  * CAS operations may fail requiring the delay calculation to be redone

Because the lock was only gaurding the section which we are not guarding with cas operations, it's not obvious to me that this is better or not.  We wont be able to get more parallism because of that.  So it really comes down to what is cheaper, synchronizing, or going to native to convert to a double to a long.  If there is high thread contention, I would be suprised if this is better.