PriorityScheduler: remove the queue manager thread
This is a large change that is really the sum of months of work.
This removes the queue manager thread.  Workers never move into an idle structure.  Instead each worker tries to agressively take tasks.  A lot more similar to how java.util's schedulers work, but with quite a few complexities due to how threadly works.
This provides a massive performance improvement.  But also still has some problems I am yet to figure out solutions for (documented as TODO).