Fix for regression in PriorityScheduler
Right before I merged the queue manager thread removal, I removed a lock when dequing tasks.  That change was ultimately incorrect.
Because a worker may null it's reference at the same time that another worker is setting that reference to be the next link in the idle chain, it means a worker may end up being lost from the chain.
This defaults back to locking during that case.  It's hard to do without locking since we must update two references in a thread safe way.