Added "ExecutorStatisticWrapper" as an ability to get statistics from any Executor
This is in part progress towards #183.  We still need an implementation that can do scheduling.
Thus this is not a complete replacement of PrioritySchedulerStatisticTracker yet.  In addition to having less functionality, it has some different performance considerations I have not fully decided on yet.
For executors those performance implications have to be swallowed no matter what.  The goal is that wrappers may be able to take advantage of threadly pool features to avoid additional clock invocations.