This resolves #192
I investigated this heavily to understand what is going on.  The condition is that several tasks are scheduled on a multi-threaded pool.
At some point a task must queue for execution.  But before it actually queues, the running task finishes.  Another task (submitted after the one that queues) starts and sees it can execute directly, so it does so, thus skipping ahead of the queued task.

I don't see an easy way to fix this, even with synchronization.  It also seems like a fairly minor problem (till someone else sees it and explains why it is not).  So for now I am just changing the unit tests back so this wont happen.