Added UnfairExecutor as a high performance specialty executor
This resolves #194 by providing an executor with superior performance, but reduced guarantees.

It distributes tasks to different queues to reduce thread contention and help solve issues where the task queue may become the bottleneck.  The default distrubtion uses the identity hash XOR'ed to the last known time.

As currently impelemted this provides around a 200% to 1000% improvement compared to java.util.concurrent.ThreadPoolExecutor.  The more cores in the system, the more this has to gain because of reduced context switches and kernel time.