Unit test minor improvements
Switching from "TestRunnable" to DoNothingRunnable in a few more places where possible without any logic changes.

Also used UnfairExecutor as backing executor in a couple places instead of using PriorityScheduler EVERYWHERE.  This helps give UnfairExecutor (which is newer) a bit more usage so we can be more confident in it.