Don't cancel futures when removed from scheduler
Weather removed from a .remove(Runnable) invokcation, or from a shutdownNow(), we don't want to cancel futures.

If we were going to automatically cancel futures, then we would need to know if we should interrupt or not (right now we never interrupt).  Because we are already checking if the task is valid there seems to be no point in this as well.
This way a user can either cancel manually, interrupt, or even just re-submit the task.