NoThreadScheduler: Performance improvements in blockingTick
This removes the object locked on to do wait/notify.
This means that if there is nothing to run, we no longer will need to synchronize and do a .wait.  Instead we will use LockSupport directly to park the thread.
In addition if a thread is blocked, on task submission we no longer need to do a synchronize and notify, instead we will invoke an unpark.

In our other schedulers we are already using LockSupport directly for the performance gains.  Because of our conditionals on when we need to synchronize it has not been a high priority to do here, but we might as well.

This gain in NoThreadScheduler will improve SingleThreadScheduler performance when the thread starts to go idle.