Pool wrappers for intercepting tasks before submision to the pool
This allows you to intercept the tasks, allowing you to do things like MDC state copies by wrapping the tasks, or modifying the task, or even avoiding the task execution (by submitting a DoNothingRunnable as the result).

This also introduces the new "wrapper" package structure, which will have a bunch of deprecations and changes soon to follow.