package org.threadly.concurrent.statistic back to org.threadly.concurrent.statistics
I decided that "statistic" sounds too weird and "collections" already creates a precedent for ending in a plural sense (which is what I was trying to avoid, plus one character savings).
In addition I figure it probably was not worth adding all the deprecations that this created (so wanted to switch back before the 4.6.0 release).