Added 'KeyedRateLimiterExecutor'
This new executor is similar to `KeyedExecutorLimiter` in that it limits based off a key.  But it applies the principles of rate limiting from `RateLimiterExecutor`.
Allowing a task submission key to indicate what rate limit should be applied to a given task.

The biggest difference implementation wise between this and `KeyedExecutorLimiter` (or more specifically `AbstractKeyedLimiter`) is in how we check if the limiter has become idle.  In `AbstractKeyedLimiter` we can look for how many tasks are running (since it's limiting concurrency).  Here instead we schedule a task out after construction.  It will run regularlly and check if the limiter is idle, if so it will remove it.  If not it will go ahead and reschedule itself to check later.  As far as if it's better or worse, it depends on how many keys there are.  For few but heavily used keys this design will be better (reduced overhead for individual task execution).

I also went through and described in the javadocs more about what "expectedParallelism" actually means.