Added ListenableFuture.map functionality
This is a huge feature that we want for 5.0.0, and just the start of threadly monads.
This map functionality can be invoked on any ListenableFuture to transform the result in an efficent way, making the code needed to work with futures much smaller and cleaner.
Included with this is some restructuring of the future unit tests to reduce code duplication.