RateLimiterExecutor: Add ability to specify maximum schedule delay
This seems like a common desire, to not want this to schedule out infinitely far, and at a certain point to just start to reject tasks.
I have been doing this behavior external to this class for a while now, but it will just be more efficent to do it in the native implementation (and with VERY minimal impacts if this feature is not desired, though I would argue that in almost every case even a very long maximum should be set).

Because this adds the functionality to `RateLimiterExecutor` we also added it to `KeyedRateLimiterExecutor`.