Limiter wrappers: Allow rejected tasks to be handled
Rather than only allowing `RejectedExecutionException`'s to be thrown, they can now be handled in other ways (or even just ignored).
The default is still to throw the `RejectedExecutionException`, but the flexibility is now there.

I looked at putting this same behavior in the normal pools, but since they only throw when shutdown, and because our constructor bloat is already really bad I decided to punt on that till I can decide what to do with the constructors (ugh, builders are starting to look better).

During this it was found the old deprecated version of `SubmitterSchedulerQueueLimitRejector` had an oversight in that it was just copied.  We can reduce some code duplication by having the old implementation extend the new one.