Extract `Watchdog`s task into an abstract `ReschedulingOperation`
The logic here is not really changing, just extracting this out to an abstract implementation so that others can take advantage of the flexible rescheduling task.

This makes it easier for tools which need to regularly check something, but don't want to have a cleanup operation for when they want to be garbage collected (ie only scheduled when something to do).

It's not really designed as a way to reduce scheduler queue load, but more focused on being flexible since while not scheduled things like schedule delay can be easily changed.