IMPORTANT 4.8.1 PriorityScheduler/SingleThreadScheduler fix
This fixes a regression introduced in 4.4.0 with PriorityScheduler and SingleThreadScheduler.

An optimization made in 4.3.0 (that was safe while the thread consumer was single threaded) ended up causing an incorrect assumption once we switched to multi-threaded queue polling in 4.4.0.

In 4.3.0 we knew we were the head of the queue from an earlier check and thus did not need to do the expense of an index lookup.  However now with parallel queue pulling we can actually pull from different parts of the queue at the same time, so even if a task scheduled ahead of the one currently ready to run (most commonly seen with scheduling at a fixed rate).

We are able to do a minor optimization to avoid the reposition by checking if we are already in an "acceptible" position.

This was found from the increased unit tests in 5.X.X branch, those tests have been migrated down to this branch as well now.