Added `Poller`
This class, backed by a scheduling pool, can take a supplier of boolean (or a java.util.concurrent Future), and return a ListenableFuture for state change.

Polling is a classic way of taking blocking or otherwise non-event driven things and making them event driven.  This class provides a simple and generic way to do this simple action.