Fix for potentially leaving waitingForUnpark to true
This fixes the following race condition:
Thread A is pool thread, it checked queue and found no work.  It has just queued itself as an idle worker
Thread B is submitting a task and notices an idle thread it needs to wake up, it unparks, and then sets the waitingForUnpark boolean
Thread A then finds something to do, and never parks, exiting workerIdle and leaving the waitingForUnpark in a bad state.

The simplest option would be to just set waitingForPark to false in the finally if we ever queued.  In this fix we track if we parked to try and continue to get maximum benefit of this optimization (assuming the management of this extra boolean does not outweigh it)