We don't need to unset the waiting for unpark here
Now that the park status is being tracked on a per-worker basis, resetting this state in the finally is no longer necessary.
In theory a thread could go and do application work, come back and need to park/queue, and still take advantage of an unused unpark call from before.