5.1 PriorityScheduler hot fix for threads not unparking
An optimization introduced in 5.0 can interact with user applications that use `LockSupport` to park/unpark.  This can happen just simple things like blocking on a future, or using a ReentrantLock.
The solution is just to reset the state once we know the thread is within our control before we park.  The best place for this seems to before we queue the thread as parked.

Lots of benchmarking occured in this change since this does reduce the original benefit.  In fact the advantage to no optimization at all is minor at best.  I suspect this might be a place for further optimization in the future.  Right now it does make sense to reduce the unparks due to the seperate peek / remove time (don't want threads wasting time inspecting tasks they can't execute).  A global check shows benefit in some use cases but worst in others.  In contrast an aggressive unpark (like searching for a parked thread) also shows benefit in some cases more than others.  This current solution is currently designed to be a safe compromise.