SchedulerService added getWaitingForExecutionTaskCount()
PrioritySchedulerService also got getWaitingForExecutionTaskCount(TaskPriority).
This is a slightly computationally more expensive call (because of the need to traverse the head of the schedule queue to find how many are waiting, as well as the possibility of an interior thrown exception if modified during traversal).  But can be used to understand more accurately what the back pressure of the thread pool truly is.